move: COMMANDS, QUERY_LIST, dispatchQuery, displayUserChoices, quit
